[refactor] MatchFinishViewクラスのstaticmethodをMatchFinishServiceクラスに切り出し
